Move config to __init__ and change import machinery
